basic/naive signal handling
The goal here was to hook Ctrl-C and send USR2 signal to childred instead.
Turns out this is doesn't work easily because Ctrl-C is sent to the entire
process group, not just the top running process. Making this work would require
either masking SIGINT in the childred (children might change this) or mucking
with process groups, both of which are "meh".

But, unsafe code, woooo!